PR suggestion implementation
naming changed to rambo.h and massless_rambo.h
flag instead of switch (--rambo-massless)
