+++ madgraph/iolibs/template_files/madmatrix/check_sa.cc
@@ -76,14 +76,14 @@ namespace
 
   enum Mode { MODE_MATRIX, MODE_PERF };
 
-  enum RamboType { RAMBO_CLASSIC, RAMBO_MASSLESS };
+  enum RamboType { RAMBO_MASSIVE, RAMBO_MASSLESS };
 
   int usage( const char* argv0, int ret = 1 )
   {
     std::cout
       << "Usage:\n"
       << "  " << argv0 << " [matrix] [-v|--verbose]\n"
-      << "  " << argv0 << " perf [-v|--verbose] [-f|--flavor <int>] [-r|--rambo [c]|ml]"
+      << "  " << argv0 << " perf [-v|--verbose] [-f|--flavor <int>] [--rambo-massless]"
       << " [<#blocksPerGrid> <#threadsPerBlock>] <#iterations>\n"
       << "  " << argv0 << " -p [opts]   (legacy alias for `perf`)\n"
       << "\n"
@@ -91,7 +91,6 @@ namespace
       << "  matrix (default)  Run " << kMatrixBlocks << " events on every flavor combination\n"
       << "                    and print the first event's phase-space point and\n"
       << "                    matrix element for each flavor.\n"
-      << "                    Uses the classic (massive) RAMBO.\n"
       << "                    With -v also prints backend/fptype/hardcodePARAM header.\n"
       << "  perf              Run #blocks*#threads events over #iterations iterations\n"
       << "                    on a single flavor index, then print performance counters.\n"
@@ -303,7 +302,7 @@ namespace
 
     // Always massive RAMBO
     std::unique_ptr<SamplingKernelBase> prsk(
-      new ClassicRamboSamplingKernelHost( kEnergy, hstRndmom, masses, CPPProcess::npari, nevt, hstMomenta, hstWeights ) );
+      new RamboSamplingKernelHost( kEnergy, hstRndmom, masses, CPPProcess::npari, nevt, hstMomenta, hstWeights ) );
     prsk->getMomentaInitial();
     prsk->getMomentaFinal();
 
@@ -417,7 +416,7 @@ namespace
 
     // Retrieve masses
     std::vector<fptype> masses;
-    if( ramboType == RAMBO_CLASSIC )
+    if( ramboType != RAMBO_MASSLESS)
     {
       int npar_meta = 0;
       if( umami_get_meta( UMAMI_META_PARTICLE_COUNT, &npar_meta ) != UMAMI_SUCCESS || npar_meta != CPPProcess::npar )
@@ -437,17 +436,17 @@ namespace
     }
 
     std::unique_ptr<SamplingKernelBase> prsk;
-    if( ramboType == RAMBO_CLASSIC )
+    if( ramboType != RAMBO_MASSLESS )
     {
       // Massive host only (copy) 
-      prsk.reset( new ClassicRamboSamplingKernelHost( kEnergy, hstRndmom, masses, CPPProcess::npari, nevt, hstMomenta, hstWeights ) );
+      prsk.reset( new RamboSamplingKernelHost( kEnergy, hstRndmom, masses, CPPProcess::npari, nevt, hstMomenta, hstWeights ) );
     }
     else
     {
 #ifdef MGONGPUCPP_GPUIMPL
-      prsk.reset( new RamboSamplingKernelDevice( kEnergy, devRndmom, devMomenta, devWeights, gpublocks, gputhreads ) );
+      prsk.reset( new MasslessRamboSamplingKernelDevice( kEnergy, devRndmom, devMomenta, devWeights, gpublocks, gputhreads ) );
 #else
-      prsk.reset( new RamboSamplingKernelHost( kEnergy, hstRndmom, hstMomenta, hstWeights, nevt ) );
+      prsk.reset( new MasslessRamboSamplingKernelHost( kEnergy, hstRndmom, hstMomenta, hstWeights, nevt ) );
 #endif
     }
 
@@ -490,15 +489,15 @@ namespace
       rambtime += timermap.stop();
 #ifdef MGONGPUCPP_GPUIMPL
       // Massive host only (copy)
-      if( ramboType == RAMBO_CLASSIC )
+      if( ramboType != RAMBO_MASSLESS )
       {
-        timermap.start( "2x CpHTDmom" );
+        timermap.start( "2c CpHTDmom" );
         copyDeviceFromHost( devMomenta, hstMomenta );
         rambtime += timermap.stop();
       }
 #endif
 
-      timermap.start( "2c Aosoa2U " );
+      timermap.start( "2d Aosoa2U " );
 #ifdef MGONGPUCPP_GPUIMPL
       gpuLaunchKernel( aosoa_to_umami_kernel, gpublocks, gputhreads, devMomenta.data(), devUmamiMomenta.data(), (std::size_t)nevt );
       checkGpu( gpuPeekAtLastError() );
@@ -627,7 +626,7 @@ int main( int argc, char** argv )
 {
 
   Mode mode = MODE_MATRIX;
-  RamboType ramboType = RAMBO_CLASSIC; // default
+  RamboType ramboType = RAMBO_MASSIVE; // default
   bool ramboTypeSet = false;
   bool verbose = false;
   unsigned int flavorID = 0;
@@ -656,12 +655,10 @@ int main( int argc, char** argv )
       mode = MODE_PERF; // legacy alias
     else if( ( arg == "--flavor" || arg == "-f" ) && argn + 1 < argc && is_number( argv[argn + 1] ) )
       flavorID = strtoul( argv[++argn], nullptr, 0 );
-    else if( ( arg == "--rambo" || arg == "-r" ) && argn + 1 < argc )
+    else if( arg == "--rambo-massless"  )
     {
       std::string r = argv[++argn];
-      if( r == "c" ) ramboType = RAMBO_CLASSIC;
-      else if( r == "ml" ) ramboType = RAMBO_MASSLESS;
-      else return usage( argv[0] );
+      ramboType = RAMBO_MASSLESS;
       ramboTypeSet = true;
     }
     else if( is_number( argv[argn] ) && nnum < 3 )
@@ -677,7 +674,7 @@ int main( int argc, char** argv )
 
   if( mode == MODE_MATRIX )
   {
-    if( ramboTypeSet && ramboType != RAMBO_CLASSIC )
+    if( ramboType == RAMBO_MASSLESS )
     {
       std::cerr << "ERROR: matrix mode only supports the classic RAMBO (-r c)." << std::endl;
       return usage( argv[0] );
